requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
chromadb>=0.5,<0.6
sentence-transformers>=3.0,<4.0
numpy>=1.26,<3.0
openai>=1.30,<2.0
